add option to provide path in settings
