+++ src/companion.ts
@@ -16,7 +16,6 @@ import {
 import { getProblemName } from './submit';
 import { spawn } from 'child_process';
 import { getJudgeViewProvider } from './extension';
-import * as customEnvironmentVariables from './customEnvironmentVariables';
 
 const emptyResponse: CphEmptyResponse = { empty: true };
 let savedResponse: CphEmptyResponse | CphSubmitResponse = emptyResponse;
@@ -145,6 +144,28 @@ export const getProblemFileName = (problem: Problem, ext: string) => {
     }
 };
 
+const getInitialText = (): string => {
+    let date_ob = new Date();
+
+    // current date
+    // adjust 0 before single digit date
+    let date = ("0" + date_ob.getDate()).slice(-2);
+
+    // current month
+    let month_name = date_ob.toLocaleString('default', { month: 'long' });
+
+    // current year
+    let year = date_ob.getFullYear();
+
+    let today_string = date + "-" + month_name + "-" + year + " " + date_ob.toLocaleTimeString('en-US', { hour12: false });
+
+    let str = '';
+    str += '\n';
+    str += '//\tparsed : ' + today_string + ' IST';
+
+    return str;
+}
+
 /** Handle the `problem` sent by Competitive Companion, such as showing the webview, opening an editor, managing layout etc. */
 const handleNewProblem = async (problem: Problem) => {
     // If webview may be focused, close it, to prevent layout bug.
@@ -200,7 +221,7 @@ const handleNewProblem = async (problem: Problem) => {
         id: randomId(),
     }));
     if (!existsSync(srcPath)) {
-        writeFileSync(srcPath, customEnvironmentVariables.getInitialText());
+        writeFileSync(srcPath, getInitialText());
     }
     saveProblem(srcPath, problem);
     const doc = await vscode.workspace.openTextDocument(srcPath);
